feat(unfurl): fetch only HTTPS images
Icebear will no longer unfurl URLs that don't start with https://

When it encounters an insecure URL which looks like it may point to an
image (ends with .jpg, .png, etc) it still adds the link to external
images without fetching it, but marks it as isInsecure. Clients should
check isInsecure in external image data and display a warning
("not showing insecure image...") instead of the actual image.